backend/worker_requirement.txt
sqlalchemy
psycopg2-binary
boto3
python-dotenv
pytesseract
Pillow